pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devils-advocate"
version = "0.1.0"
description = "Cost-aware multi-LLM adversarial review engine with deterministic governance"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
authors = [{name = "Brian Kelley"}]
keywords = ["llm", "code-review", "governance", "adversarial", "consensus"]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich[markdown]>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "httpx>=0.27",
]
e2e = [
    "pytest-playwright>=0.5",
    "playwright>=1.48",
    "Pillow>=10.0",
    "pixelmatch>=0.3",
]

[project.scripts]
dvad = "devils_advocate.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: e2e tests that make real API calls (select with '-m live')",
    "e2e: end-to-end Playwright GUI tests (select with '-m e2e')",
    "e2e_live: E2E tests requiring local LLM (select with '-m e2e')",
]

[tool.hatch.build.targets.wheel]
packages = ["src/devils_advocate"]
include = [
    "src/devils_advocate/templates/*.txt",
    "src/devils_advocate/gui/static/*",
    "src/devils_advocate/gui/templates/*",
    "src/devils_advocate/examples/*",
]
